+++ system-tests/gradle-multi-module/gradle-project/app/src/main/java/com/example/app/Main.java
@@ -1,0 +1,10 @@
+package com.example.app;
+
+import com.example.lib.Calculator;
+
+public class Main {
+	public static void main(String[] args) {
+		Calculator cal = new Calculator();
+		System.out.println("4/2+3 = " + cal.add(cal.div(4, 2), 3));
+	}
+}