+++ system-tests/gradle-multi-module/gradle-project/module1/src/main/java/org/example/Main1.java
@@ -1,7 +1,0 @@
-package org.example;
-
-public class Main1 {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}