+++ system-tests/gradle-multi-module/gradle-project/module2/src/main/java/org/example/Main2.java
@@ -1,7 +1,0 @@
-package org.example;
-
-public class Main2 {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}