+++ teamscale-gradle-plugin/src/main/kotlin/com/teamscale/config/extension/TeamscaleTestImpactedTaskExtension.kt
@@ -1,46 +1,30 @@
 package com.teamscale.config.extension
 
-import com.teamscale.TestImpacted
 import com.teamscale.config.AgentConfiguration
-import com.teamscale.config.TestwiseCoverageConfiguration
-import groovy.lang.Closure
 import org.gradle.api.Action
-import org.gradle.api.Project
+import org.gradle.api.file.FileCollection
+import org.gradle.api.model.ObjectFactory
+import org.gradle.kotlin.dsl.newInstance
 import org.gradle.testing.jacoco.plugins.JacocoTaskExtension
 import java.io.Serializable
+import javax.inject.Inject
 
 /**
  * Holds all user configuration regarding testwise coverage report uploads.
  */
-open class TeamscaleTestImpactedTaskExtension(
-    val project: Project,
-    jacocoExtension: JacocoTaskExtension,
-    testImpactedTask: TestImpacted
+@Suppress("unused")
+open class TeamscaleTestImpactedTaskExtension @Inject constructor(
+    objects: ObjectFactory,
+    teamscaleJaCoCoAgentConfiguration: FileCollection,
+    jacocoExtension: JacocoTaskExtension
 ) : Serializable {
 
-    val report = TestwiseCoverageConfiguration(project, testImpactedTask)
-
-    /** Configures the reports to be uploaded. */
-    fun report(action: Action<in TestwiseCoverageConfiguration>) {
-        action.execute(report)
-    }
-
-    /** Overload for Groovy DSL compatibility. */
-    fun report(closure: Closure<*>) {
-        report { project.configure(this, closure) }
-    }
-
     /** Settings regarding the teamscale jacoco agent. */
-    val agent = AgentConfiguration(project, jacocoExtension)
+    val agent = objects.newInstance<AgentConfiguration>(teamscaleJaCoCoAgentConfiguration, jacocoExtension)
 
     /** Configures the jacoco agent options. */
     fun agent(action: Action<in AgentConfiguration>) {
         action.execute(agent)
-    }
-
-    /** Overload for Groovy DSL compatibility. */
-    fun agent(closure: Closure<*>) {
-        agent { project.configure(this, closure) }
     }
 }