+++ teamscale-gradle-plugin/test-project/src/main/java/com/example/project/Fibonacci.java
@@ -1,13 +1,15 @@
+package com.example.project;
+
 public class Fibonacci {
     public static int compute(int n) {
     	int result = 0;
-    	
-        if (n <= 1) { 
-        	result = n; 
-        } else { 
-        	result = compute(n - 1) + compute(n - 2); 
+
+        if (n <= 1) {
+        	result = n;
+        } else {
+        	result = compute(n - 1) + compute(n - 2);
         }
-        
+
         return result;
     }
 }