+++ teamscale-gradle-plugin/test-project/src/test/java/com/example/project/FibonacciTest.java
@@ -1,3 +1,5 @@
+package com.example.project;
+
 import static org.junit.Assert.assertEquals;
 
 import java.util.Arrays;
@@ -12,8 +14,8 @@ import org.junit.runners.Parameterized.Parameters;
 public class FibonacciTest {
     @Parameters
     public static Collection<Object[]> data() {
-        return Arrays.asList(new Object[][] {     
-                 { 0, 0 }, { 1, 1 }, { 2, 1 }, { 3, 2 }, { 4, 3 }, { 5, 5 }, { 6, 8 }  
+        return Arrays.asList(new Object[][] {
+                 { 0, 0 }, { 1, 1 }, { 2, 1 }, { 3, 2 }, { 4, 3 }, { 5, 5 }, { 6, 8 }
            });
     }